Bug fixes - Fixed an issue with the calendar, now I store the date as year/month/day in the database directly. - VB expects to have a year between 2021 and 2032, for some reason it gives me back a 2000 year when reading an empty history. I just patched it in the application to return as year 0 in those cases. - VitalsHistory works as intended now. Also verified functionality of the new classes. - Moved a few SQL queries too, since they didn't belong where they were.
Also added a small feature, now the application knows if you entered the scan screen through a character in the storage, or the home screen. If there is an active character and you opened the home screen, the application will let you send the active character, while if it was opened through the storage screen, the application will send the character from the storage screen.

What's now missing is to create a VBSCanScreen, and depending on which device we are using, make use of one screen or the other.